requirements.txt
asgiref==3.11.0 ; python_version >= "3.12" and python_version < "4.0"
certifi==2025.11.12 ; python_version >= "3.12" and python_version < "4.0"
cffi==2.0.0 ; python_version >= "3.12" and python_version < "4.0" and platform_python_implementation != "PyPy"
charset-normalizer==3.4.4 ; python_version >= "3.12" and python_version < "4.0"
click==8.3.1 ; python_version >= "3.12" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and platform_system == "Windows"
cryptography==46.0.3 ; python_version >= "3.12" and python_version < "4.0"
django-allauth==65.13.1 ; python_version >= "3.12" and python_version < "4.0"
django==6.0 ; python_version >= "3.12" and python_version < "4.0"
h11==0.16.0 ; python_version >= "3.12" and python_version < "4.0"
idna==3.11 ; python_version >= "3.12" and python_version < "4.0"
psycopg2-binary==2.9.11 ; python_version >= "3.12" and python_version < "4.0"
pycparser==2.23 ; python_version >= "3.12" and python_version < "4.0" and platform_python_implementation != "PyPy" and implementation_name != "PyPy"
pyjwt==2.10.1 ; python_version >= "3.12" and python_version < "4.0"
python-dotenv==1.2.1 ; python_version >= "3.12" and python_version < "4.0"
requests==2.32.5 ; python_version >= "3.12" and python_version < "4.0"
sqlparse==0.5.5 ; python_version >= "3.12" and python_version < "4.0"
tzdata==2025.3 ; python_version >= "3.12" and python_version < "4.0" and sys_platform == "win32"
urllib3==2.6.2 ; python_version >= "3.12" and python_version < "4.0"
uvicorn==0.40.0 ; python_version >= "3.12" and python_version < "4.0"
